ADD:增加代理功能
