+++ lib/requests/requests.py
@@ -1,5 +1,10 @@
 import httpx
 import json
+import time
+from data.driver import Proxies
+            
+proxyModel = Proxies("data/proxies/proxies.db")
+proxies = []
 
 class Response:
     def __init__(self, status_code, text):
@@ -21,14 +26,24 @@ def retry_request(func, max_retries=3):
 
     return wrapper
 
+async def get_proxy():
+    global proxies
+    proxies = await proxyModel.load()
+    if(len(proxies) == 0):
+        return None
+    proxy = proxies[int(int(time.time()) / 300) % len(proxies)]
+    return proxy['url']
+
 @retry_request
 async def get(url, headers=None, params=None) -> Response:
-    async with httpx.AsyncClient() as client:
+    proxy = await get_proxy()
+    async with httpx.AsyncClient(proxy=proxy) as client:
         response = await client.get(url, headers=headers, params=params)
         return Response(response.status_code, response.text)
 
 @retry_request
 async def post(url, headers=None, data=None, json=None) -> Response:
-    async with httpx.AsyncClient() as client:
+    proxy = await get_proxy()
+    async with httpx.AsyncClient(proxy=proxy) as client:
         response = await client.post(url, headers=headers, json=json, data=data)
         return Response(response.status_code, response.text)
